+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Methods throw a `SocketException` when used in an invalid state
 - Exposed a `clients` getter in `Server` class to access the connected clients
 - Replaced the default server port with a random port chosen by the system
+- Removed `remotePort` in `Client` class as it's already known when connecting to the server
 
 ## 1.0.1